backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.6.0
pydantic-settings>=2.1.0
anthropic>=0.18.0
pillow>=10.2.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=23.2.1
